LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/animations/font-palette-animation-not-specified-endpoints-expected.txt

A
A

PASS Verify font-palette is animated when `from` keyframe is not specified
FAIL Verify font-palette is animated when `to` keyframe is not specified assert_equals: expected "normal" but got "--custom"
